feat(parser) Factorize `block_void` and `block_balanced`.
A `block` is defined as a disjunction of a `block_void` or a
`block_balanced`. They both start with the same pattern. If one fails,
then the same data will be parsed the same exact way. This is a waste
of time, and code duplication.

The new `block` definition factorizes the beginning of both block
definitions, and sets the disjunction for the end of the patterns.

This patch reduces the WASM file size by 3Kb.