fix(php) Fix all memory leaks.
The `efree` call in the object destructor was not required. And more
importantly, while it fixed a first memory leak, when calling
`gutenberg_post_parser($content);` multiple times, a segfault was
fired.

The 2 other memory leaks have been solved by decrementing the GC
counter for children in a block with `Z_DELREF`. And it solves the
first memory leak at the same time.

All memory leaks reported by `--enable-debug` are solved.