feat(c) Create `Slice_c_char` to avoid memory copies with `CString`.
`CString` copies the data on the heap, and we don't want to copy the
data. So instead, we add a new struct `Slice_c_char` which holds a
pointer and a length. The pointer points “inside” the given input, so
freeing memory is easier in this scenario: A single data to free.