feat(php) Avoid memory copies by using `Slice_c_char`.
The C binding exposes a `Slice_c_char` instead of a `CString`
pointer. Thus there is one less copy in memory.

Before `Slice_c_char`, the data were copied in memory a first time
from the input to `Cstring`, then a second time from the `CString` to
`zend_string`. Since Zend Engine 3, all `zend_string` are
automatically copied in memory. So to remove one copy, the `CString`
API has been removed, and replaced by `Slice_c_char`. Hopefully, PHP
provides a `ZVAL_STRINGL` macro to specify the length of the
string. The string can not end by a NUL byte because if the length is
specified then it's considered as binary safe.

Remember the definition of `zend_string` at the time of writing this
commit:

```c
struct _zend_string {
	zend_refcounted_h gc;
	zend_ulong        h;                /* hash value */
	size_t            len;
	char              val[1];
};
```