feat(wasm) Decode block's name & attributes in one go.
Instead of calling the `TextDecoder` once for name, and once for
attributes (when there is some), we call the `TextDecoder` only once,
and then split the string to retrieve the name and the attributes. It
works because the name is only in the range `[a-z0-9\-]`, so it maps
`u8`.

It saves 0.5% of the CPU.