feat(wasm) Use `subarray` instead of `slice` to avoid copy.
Even if `slice` promises a shallow copy, this is not for all
VM. `subarray` shares the same array buffer than the source, so there
is zero copy. It boosts performance by 16%.

Also, this patch uses `Uint8ClampedArray` instead of `Uint8Array`. I
hoped it could boost the performance a little bit but there is no
noticeable difference.