feat(wasm) Push UTF-16 offsets in memory instead of u8 offsets.
The entire datum is encoded at &[u8] in the WASM memory.

Before this patch, the `root` function was returning offsets to the
memory, and the JavaScript boundary had to decode the datum from &[u8]
to JS strings (encoded in UTF-16) with the help of `TextDecoder`.

`TextDecoder` is a performance bottleneck. We try to get rid of it.

After this patch, the `root` function returns UTF-16 offsets, so that
the JavaScript boundary can substrings directly on the datum without
reading the WASM boundary again, and thus without using the
`TextDecoder` API.